+++ src/main/java/edu/classworks/JavaOnWeb/services/StudentService.java
@@ -21,7 +21,7 @@ public class StudentService extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String action = request.getParameter("action");
-        int studentId = Integer.parseInt(request.getParameter("id"));
+        long studentId = Long.parseLong(request.getParameter("id"));
 
         switch (action) {
             case "DELETE":
